moved util.hpp in separate folder
